rename:poresonal-info index 추가 및 파일이동(#387)
